Move access checker and @AnonymousAllowed to flow-server
This allows them to be used for view access control also
