+++ fusion-endpoint/src/main/java/com/vaadin/flow/server/connect/VaadinConnectControllerConfiguration.java
@@ -18,16 +18,16 @@ package com.vaadin.flow.server.connect;
 
 import java.lang.reflect.Method;
 
+import com.vaadin.flow.server.auth.AccessAnnotationChecker;
+import com.vaadin.flow.server.connect.auth.CsrfChecker;
+import com.vaadin.flow.server.connect.auth.VaadinConnectAccessChecker;
+
 import org.springframework.boot.autoconfigure.web.servlet.WebMvcRegistrations;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.mvc.condition.PatternsRequestCondition;
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
 import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
-
-import com.vaadin.flow.server.connect.auth.AccessAnnotationChecker;
-import com.vaadin.flow.server.connect.auth.CsrfChecker;
-import com.vaadin.flow.server.connect.auth.VaadinConnectAccessChecker;
 
 /**
  * A configuration class for customizing the {@link VaadinConnectController}
