+++ fusion-endpoint/src/test/java/com/vaadin/flow/server/connect/generator/endpoints/denyall/DenyAllEndpoint.java
@@ -20,8 +20,8 @@ import javax.annotation.security.DenyAll;
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
 
+import com.vaadin.flow.server.auth.AnonymousAllowed;
 import com.vaadin.flow.server.connect.Endpoint;
-import com.vaadin.flow.server.connect.auth.AnonymousAllowed;
 
 @Endpoint
 @DenyAll
