+++ fusion-endpoint/src/test/java/com/vaadin/flow/server/connect/generator/endpoints/json/JsonTestEndpoint.java
@@ -15,20 +15,19 @@
  */
 package com.vaadin.flow.server.connect.generator.endpoints.json;
 
-import javax.annotation.security.DenyAll;
-import javax.annotation.security.PermitAll;
-import javax.annotation.security.RolesAllowed;
-
 import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import com.fasterxml.jackson.core.Version;
+import javax.annotation.security.DenyAll;
+import javax.annotation.security.PermitAll;
+import javax.annotation.security.RolesAllowed;
 
+import com.fasterxml.jackson.core.Version;
+import com.vaadin.flow.server.auth.AnonymousAllowed;
 import com.vaadin.flow.server.connect.Endpoint;
-import com.vaadin.flow.server.connect.auth.AnonymousAllowed;
 
 /**
  * This class is used for OpenApi generator test
